requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
pdfplumber
torch
pydantic
accelerate
bitsandbytes
